fix ConverterStrategyBaseLibTest. Refactoring of BalancerComposableStableUniversalTest - move common functions and types to util-classes
