+++ test/strategies/polygon/balancer/BalancerComposableStableUniversalTest.ts
@@ -33,6 +33,8 @@ import {writeFileSync} from "fs";
 import {formatUnits} from "ethers/lib/utils";
 import hre, {ethers} from "hardhat";
 import {IUniversalStrategyInputParams} from "../../base/UniversalStrategyTest";
+import {IState, UniversalTestUtils} from "../../../baseUT/utils/UniversalTestUtils";
+import {BalancerIntTestUtils} from "./utils/BalancerIntTestUtils";
 
 dotEnvConfig();
 // tslint:disable-next-line:no-var-requires
@@ -52,272 +54,6 @@ const argv = require('yargs/yargs')()
 // const {expect} = chai;
 chai.use(chaiAsPromised);
 
-//region Utils
-interface IState {
-  title: string;
-  block: number;
-  blockTimestamp: number;
-  signer: {
-    usdc: BigNumber;
-  }
-  user: {
-    usdc: BigNumber;
-  }
-  strategy: {
-    usdc: BigNumber;
-    usdt: BigNumber;
-    dai: BigNumber;
-    bal: BigNumber;
-    bptPool: BigNumber;
-    totalAssets: BigNumber;
-    investedAssets: BigNumber;
-  }
-  gauge: {
-    strategyBalance: BigNumber;
-  }
-  balancerPool: {
-    bbAmUsdc: BigNumber;
-    bbAmUsdt: BigNumber;
-    bbAmDai: BigNumber;
-  }
-  splitter: {
-    usdc: BigNumber;
-    totalAssets: BigNumber;
-  }
-  vault: {
-    usdc: BigNumber;
-    userShares: BigNumber;
-    signerShares: BigNumber;
-    userUsdc: BigNumber;
-    signerUsdc: BigNumber;
-    sharePrice: BigNumber;
-    totalSupply: BigNumber;
-    totalAssets: BigNumber;
-  },
-  insurance: {
-    usdc: BigNumber;
-  }
-}
-
-async function getStates(title: string, h: DoHardWorkLoopBase) : Promise<IState>{
-  const gauge = "0x1c514fEc643AdD86aeF0ef14F4add28cC3425306";
-  const balancerPool = "0x48e6b98ef6329f8f0a30ebb8c7c960330d648085";
-  const bbAmDai = "0x178E029173417b1F9C8bC16DCeC6f697bC323746";
-  const bbAmUsdc = "0xF93579002DBE8046c43FEfE86ec78b1112247BB8";
-  const bbAmUsdt = "0xFf4ce5AAAb5a627bf82f4A571AB1cE94Aa365eA6";
-  const splitterAddress = await h.vault.splitter();
-  const insurance = await h.vault.insurance();
-  const block = await hre.ethers.provider.getBlock("latest");
-
-  const dest = {
-    title,
-    block: block.number,
-    blockTimestamp: block.timestamp,
-    signer: {
-      usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, h.user).balanceOf(h.signer.address),
-    },
-    user: {
-      usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, h.user).balanceOf(h.user.address),
-    },
-    strategy: {
-      usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, h.user).balanceOf(h.strategy.address),
-      usdt: await IERC20__factory.connect(MaticAddresses.USDT_TOKEN, h.user).balanceOf(h.strategy.address),
-      dai: await IERC20__factory.connect(MaticAddresses.DAI_TOKEN, h.user).balanceOf(h.strategy.address),
-      bal: await IERC20__factory.connect(MaticAddresses.BAL_TOKEN, h.user).balanceOf(h.strategy.address),
-      bptPool: await IERC20__factory.connect(balancerPool, h.user).balanceOf(h.strategy.address),
-      totalAssets: await h.strategy.totalAssets(),
-      investedAssets: await StrategyBaseV2__factory.connect(h.strategy.address, h.user).investedAssets()
-    },
-    gauge: {
-      strategyBalance: await IBalancerGauge__factory.connect(gauge, h.user).balanceOf(h.strategy.address),
-    },
-    balancerPool: {
-      bbAmUsdc: await IERC20__factory.connect(bbAmUsdc, h.user).balanceOf(balancerPool),
-      bbAmUsdt: await IERC20__factory.connect(bbAmUsdt, h.user).balanceOf(balancerPool),
-      bbAmDai: await IERC20__factory.connect(bbAmDai, h.user).balanceOf(balancerPool),
-    },
-    splitter: {
-      usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, h.user).balanceOf(splitterAddress),
-      totalAssets: await ISplitter__factory.connect(splitterAddress, h.user).totalAssets(),
-    },
-    vault: {
-      usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, h.user).balanceOf(h.vault.address),
-      userShares: await h.vault.balanceOf(h.user.address),
-      signerShares: await h.vault.balanceOf(h.signer.address),
-      userUsdc: await h.vault.convertToAssets(await h.vault.balanceOf(h.user.address)),
-      signerUsdc: await h.vault.convertToAssets(await h.vault.balanceOf(h.signer.address)),
-      sharePrice: await h.vault.sharePrice(),
-      totalSupply: await h.vault.totalSupply(),
-      totalAssets: await h.vault.totalAssets(),
-    },
-    insurance: {
-      usdc: await IERC20__factory.connect(MaticAddresses.USDC_TOKEN, h.user).balanceOf(insurance),
-    },
-  }
-
-
-  console.log("State", dest);
-  return dest;
-}
-
-async function saveToCSV(pathOut: string, states: IState[]) {
-
-  const headers = [
-    "title",
-    "block",
-    "timestamp",
-    "$signer",
-    "$user",
-    "vault-$user",
-    "vault-$signer",
-    "sharePrice-vault",
-    "totalSupply-vault",
-    "totalAssets-vault",
-    "$insurance",
-    "$strategy",
-    "usdt-strategy",
-    "dai-strategy",
-    "bal-strategy",
-    "bptp-strategy",
-    "totalAssets-strategy",
-    "investedAssets-strategy",
-    "bptp-gauge",
-    "$vault",
-    "$splitter",
-    "totalAssets-splitter",
-    "bbAmUsdc-pool",
-    "bbAmUsdt-pool",
-    "bbAmDai-pool",
-  ];
-  const decimalsSharedPrice = 6;
-  const decimalsUSDC = 6;
-  const decimalsUSDT = 6;
-  const decimalsDAI = 18;
-  const decimalsBAL = 18;
-  const decimalsBbAmUsdc = 18;
-  const decimalsBbAmUsdt = 18;
-  const decimalsBbAmDai = 18;
-  const decimalsBptp = 18;
-  const decimals = [
-    0,
-    0,
-    0,
-    decimalsUSDC, // signer.usdc
-    decimalsUSDC, // user.usdc
-    decimalsUSDC, // vault.userShares
-    decimalsUSDC, // vault.signerShares
-    decimalsUSDC, // vault.userUsdc
-    decimalsUSDC, // vault.signerUsdc
-    decimalsSharedPrice, // shared price
-    decimalsUSDC, // vault.totlaSupply
-    decimalsUSDC, // vault.totalAssets
-    decimalsUSDC, // insurance.usdc
-    decimalsUSDC, // strategy.usdc
-    decimalsUSDT, // strategy.usdt
-    decimalsDAI, // strategy.dai
-    decimalsBAL, // strategy.bal
-    decimalsBptp, // strategy.bptPool
-    decimalsUSDC, // strategy.totalAssets
-    decimalsUSDC, // strategy.investedAssets
-    decimalsBptp, // gauge.strategyBalance
-    decimalsUSDC, // vault.usdc
-    decimalsUSDC, // splitter.usdc
-    decimalsUSDC, // splitter.totalAssets,
-    decimalsBbAmUsdc,
-    decimalsBbAmUsdt,
-    decimalsBbAmDai
-  ];
-  writeFileSync(pathOut, headers.join(";") + "\n", {encoding: 'utf8', flag: "a" });
-  for (const item of states) {
-    const line = [
-      item.title,
-      item.block,
-      item.blockTimestamp,
-      item.signer.usdc,
-      item.user.usdc,
-      item.vault.userShares,
-      item.vault.signerShares,
-      item.vault.userUsdc,
-      item.vault.signerUsdc,
-      item.vault.sharePrice,
-      item.vault.totalSupply,
-      item.vault.totalAssets,
-      item.insurance.usdc,
-      item.strategy.usdc,
-      item.strategy.usdt,
-      item.strategy.dai,
-      item.strategy.bal,
-      item.strategy.bptPool,
-      item.strategy.totalAssets,
-      item.strategy.investedAssets,
-      item.gauge.strategyBalance,
-      item.vault.usdc,
-      item.splitter.usdc,
-      item.splitter.totalAssets,
-      item.balancerPool.bbAmUsdc,
-      item.balancerPool.bbAmUsdt,
-      item.balancerPool.bbAmDai
-    ];
-    writeFileSync(pathOut,
-      line.map((x, index) =>
-        typeof x === "object"
-          ? +formatUnits(x, decimals[index])
-          : "" + x
-      ).join(";") + "\n",
-      {encoding: 'utf8', flag: "a"}
-    );
-  }
-}
-
-function getTotalUsdAmount(state: IState) : BigNumber {
-  return state.user.usdc.add(
-    state.signer.usdc
-  ).add(
-    state.vault.usdc
-  ).add(
-    state.insurance.usdc
-  ).add(
-    state.strategy.usdc
-  ).add(
-    state.splitter.usdc
-  );
-}
-/**
- * Get initial state marked as "enter"
- * and final state marked as "final"
- * @param states
- */
-function outputProfit(states: IState[]) {
-  if (states.length < 2) return;
-
-  const enter: IState = states[0];
-  const final: IState = states[states.length - 1];
-
-  // ethereum timestamp is in seconds
-  // https://ethereum.stackexchange.com/questions/7853/is-the-block-timestamp-value-in-solidity-seconds-or-milliseconds
-  const timeSeconds = (final.blockTimestamp - enter.blockTimestamp);
-  const initialAmount = getTotalUsdAmount((enter));
-  const finalAmount = getTotalUsdAmount(final);
-  const amount = finalAmount.sub(initialAmount);
-  const amountNum = +formatUnits(amount, 6);
-  const apr = amountNum * 365
-    / (timeSeconds / (24*60*60))
-    / +formatUnits(initialAmount, 6)
-    * 100;
-  console.log("final.blockTimestamp", final.blockTimestamp);
-  console.log("enter.blockTimestamp", enter.blockTimestamp);
-  console.log("final.getTotalUsdAmount", getTotalUsdAmount(final));
-  console.log("final.getTotalUsdAmount", getTotalUsdAmount(enter));
-  console.log("Initial amount", initialAmount);
-  console.log("Final amount", initialAmount);
-  console.log("Total profit", amountNum);
-  console.log("Duration in seconds", timeSeconds);
-  console.log("Duration in days", timeSeconds / (24*60*60));
-  console.log("Estimated APR, %", apr);
-
-}
-//endregion Utils
-
 describe('BalancerComposableStableUniversalTest', async () => {
   if (argv.disableStrategyTests || argv.hardhatChainId !== 137) {
     return;
@@ -326,51 +62,33 @@ describe('BalancerComposableStableUniversalTest', async () => {
   const deployInfo: DeployInfo = new DeployInfo();
   const states: IState[] = [];
   const core = Addresses.getCore();
+  const tetuConverterAddress = getConverterAddress();
 
   before(async function () {
     await StrategyTestUtils.deployCoreAndInit(deployInfo, argv.deployCoreContracts);
 
     const [signer] = await ethers.getSigners();
-    // set up health factors in tetu converter
-    // set min health factor 1.02
-    // for dai and usdt set target health factor = 1.05
-    const controllerAddress = await ITetuConverter__factory.connect(getConverterAddress(), signer).controller();
-    const controller = IConverterController__factory.connect(controllerAddress, signer);
-    const governance = await controller.governance();
-    const controllerAsGovernance = IConverterController__factory.connect(
-      controllerAddress,
-      await Misc.impersonate(governance)
-    );
-    const borrowManagerAddress = await controller.borrowManager();
-    await controllerAsGovernance.setMinHealthFactor2(102);
-    const borrowManagerAsGovernance = IBorrowManager__factory.connect(
-      borrowManagerAddress,
-      await Misc.impersonate(governance)
-    );
-    await controllerAsGovernance.setTargetHealthFactor2(112);
-    await borrowManagerAsGovernance.setTargetHealthFactors(
-      [MaticAddresses.USDC_TOKEN, MaticAddresses.DAI_TOKEN, MaticAddresses.USDT_TOKEN],
-      [112, 112, 112]
-    );
 
-    const splitterImpl = await DeployerUtils.deployContract(signer, 'StrategySplitterV2')
-    await VaultFactory__factory.connect(
-      core.vaultFactory,
-      await DeployerUtilsLocal.getControllerGovernance(signer)
-    ).setSplitterImpl(splitterImpl.address);
+    await BalancerIntTestUtils.setTetConverterHealthFactors(signer, tetuConverterAddress);
+    await BalancerIntTestUtils.deployAndSetCustomSplitter(signer, core);
+
+    // Disable DForce (as it reverts on repay after block advance)
+    await ConverterUtils.disablePlatformAdapter(signer, getDForcePlatformAdapter());
+
+    // Disable Hundred Finance (no liquidity)
+    await ConverterUtils.disablePlatformAdapter(signer, getHundredFinancePlatformAdapter());
   });
 
   /** Save collected states to csv, compute profit */
   after(async function() {
     const pathOut = "./tmp/ts2-snapshots.csv";
-    await saveToCSV(pathOut, states);
-    outputProfit(states);
+    await UniversalTestUtils.saveListStatesToCSV(pathOut, states);
+    UniversalTestUtils.outputProfit(states);
   });
 
   const strategyName = 'BalancerComposableStableStrategy';
   const assetName = 'USDC';
   const asset = PolygonAddresses.USDC_TOKEN;
-  const vaultName = 'tetu' + assetName;
   const params: IUniversalStrategyInputParams = {
     ppfsDecreaseAllowed: false,
     balanceTolerance:  0.000001, // looks like some rounding issues with 6-decimals tokens
@@ -385,50 +103,24 @@ describe('BalancerComposableStableUniversalTest', async () => {
     }
   }
 
-  const deployer = async (signer: SignerWithAddress) => {
-    const controller = DeployerUtilsLocal.getController(signer);
-
-    const strategyDeployer = async (splitterAddress: string) => {
-      const strategy = BalancerComposableStableStrategy__factory.connect(
-        await DeployerUtils.deployProxy(signer, strategyName),
-        signer
-      );
-
-      await strategy.init(
-        core.controller,
-        splitterAddress,
-        getConverterAddress(),
-      );
-
-      // Disable DForce (as it reverts on repay after block advance)
-      await ConverterUtils.disablePlatformAdapter(signer, getDForcePlatformAdapter());
-
-      // Disable Hundred Finance (no liquidity)
-      await ConverterUtils.disablePlatformAdapter(signer, getHundredFinancePlatformAdapter());
-
-      return strategy as unknown as IStrategyV2;
-    }
-
-    console.log('getControllerGovernance...');
-    const gov = await DeployerUtilsLocal.getControllerGovernance(signer);
-
-    console.log('deployAndInitVaultAndStrategy...');
-    return DeployerUtilsLocal.deployAndInitVaultAndStrategy(
-      asset, vaultName, strategyDeployer, controller, gov,
-      100, 250, 500, false
-    );
-  }
-
   /* tslint:disable:no-floating-promises */
   await startDefaultStrategyTest(
     strategyName,
     asset,
     assetName,
     deployInfo,
-    deployer,
+    await UniversalTestUtils.makeStrategyDeployer(
+      core,
+      asset,
+      tetuConverterAddress,
+      {
+        strategyName,
+        vaultName: 'tetu' + assetName
+      }
+    ),
     params,
     async (title, h) => {
-      states.push(await getStates(title, h));
+      states.push(await UniversalTestUtils.getStates(title, h));
     }
   );
 });
